ended integration with statistics components and web service plus some tests

+++ game-backend/src/main/java/com/games/game/results/controller/ResultController.java
@@ -4,7 +4,7 @@ import com.games.game.move.models.MoveDto;
 import com.games.game.results.models.MoveStatsDto;
 
 public interface ResultController {
-    public Float putPlayerWinStatistics ();
+    public Integer putPlayerWinStatistics ();
     public MoveStatsDto getMoveStatistics (String move);
 
 }
