+++ game-backend/src/main/java/com/games/game/results/service/ResultService.java
@@ -4,7 +4,7 @@ import com.games.game.move.models.MoveTypeEnum;
 import com.games.game.results.models.MoveStatsDto;
 
 public interface ResultService {
-    public Float playerWinStatistics ();
+    public Integer playerWinStatistics ();
     public MoveStatsDto moveStatistics (String move);
 
 }
